# packages/api-contracts/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: Actenora Platform API
  version: 0.2.0
  description: Platform + web-portal contract surface for the modular monolith (FAZ 24).
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/meetings:
    get:
      operationId: listMeetings
      summary: Cursor-paginated meeting list
      parameters:
        - $ref: "#/components/parameters/StatusFilter"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Meeting page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingListResponse"
  /api/v1/meetings/{meetingId}:
    get:
      operationId: getMeetingDetail
      summary: Full meeting detail for three-panel view
      parameters:
        - $ref: "#/components/parameters/MeetingId"
      responses:
        "200":
          description: Meeting detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingDetailResponse"
...
components:
  parameters:
    MeetingId:
      name: meetingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    StatusFilter:
      name: status
      in: query
      schema:
        type: string
  schemas:
...
    MeetingOccurrenceStatus:
      type: string
      enum: [SCHEDULED, IN_PROGRESS, ENDED, CANCELLED, PROCESSING, READY, FAILED]
    MeetingSummary:
      type: object
      additionalProperties: false
      required: [id, title, status, scheduledStartAt, participantCount]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          $ref: "#/components/schemas/MeetingOccurrenceStatus"
        scheduledStartAt:
          type: string
          format: date-time
        participantCount:
          type: integer
    MeetingListResponse:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MeetingSummary"
        nextCursor:
          type: string
          nullable: true
    ArtifactStatus:
      type: string
      enum: [DRAFT, PENDING_APPROVAL, APPROVED, REJECTED, COMPLETED, OPEN, AT RISK]
    EvidenceRef:
      type: object
      additionalProperties: false
      required: [segmentId, startMs, endMs, quote]
      properties:
        segmentId:
          type: string
          format: uuid
        startMs:
          type: integer
        endMs:
          type: integer
        quote:
          type: string
    MeetingNote:
      type: object
      additionalProperties: false
      required: [id, visibility, body, updatedAt, authorId]
      properties:
        id:
          type: string
          format: uuid
        visibility:
          type: string
          enum: [SHARED, PRIVATE]
        body:
          type: string
        updatedAt:
          type: string
          format: date-time
        authorId:
          type: string
          format: uuid
...
    DecisionItem:
      type: object
      additionalProperties: false
      required: [id, meetingId, title, status, evidence, createdAt]
      properties:
        id:
          type: string
          format: uuid
        meetingId:
          type: string
          format: uuid
        title:
          type: string
        status:
          $ref: "#/components/schemas/ArtifactStatus"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceRef"
        createdAt:
          type: string
          format: date-time
    ActionItem:
      type: object
      additionalProperties: false
      required: [id, meetingId, title, status, ownerDisplayName, dueAt, evidence]
      properties:
        id:
          type: string
          format: uuid
        meetingId:
          type: string
          format: uuid
        title:
          type: string
        status:
          $ref: "#/components/schemas/ArtifactStatus"
        ownerDisplayName:
          type: string
        dueAt:
          type: string
          format: date-time
          nullable: true
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceRef"
    RiskItem:
      type: object
      additionalProperties: false
      required: [id, title, severity, evidence]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceRef"
    CommitmentItem:
      type: object
      additionalProperties: false
      required: [id, meetingId, statement, ownerDisplayName, dueAt, status, evidence]
      properties:
        id:
          type: string
          format: uuid
        meetingId:
          type: string
          format: uuid
        statement:
          type: string
        ownerDisplayName:
          type: string
        dueAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: "#/components/schemas/ArtifactStatus"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceRef"
    ApprovalRecord:
      type: object
      additionalProperties: false
      required: [id, artifactType, artifactId, status]
      properties:
        id:
          type: string
          format: uuid
        artifactType:
          type: string
        artifactId:
          type: string
          format: uuid
        noteId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [PENDING, GRANTED, DENIED, EXPIRED, CHANGES_REQUESTED]
        version:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
          nullable: true
        decidedBy:
          type: string
          nullable: true
        decidedAt:
          type: string
          format: date-time
          nullable: true
        comment:
          type: string
          nullable: true
...
    Participant:
      type: object
      additionalProperties: false
      required: [id, displayName, email, participantType, attendanceStatus, external]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        email:
          type: string
        participantType:
          type: string
        attendanceStatus:
          type: string
        external:
          type: boolean
    MeetingVersion:
      type: object
      additionalProperties: false
      required: [version, label, createdAt]
      properties:
        version:
          type: integer
        label:
          type: string
        createdAt:
          type: string
          format: date-time
    MeetingDetailResponse:
      type: object
      additionalProperties: false
      required:
        - meeting
        - participants
        - seriesTitle
        - businessContext
        - versions
        - approvalHistory
        - notes
        - decisions
        - actions
        - risks
        - commitments
        - qualityFlags
        - partial
      properties:
        meeting:
          $ref: "#/components/schemas/MeetingSummary"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
        seriesTitle:
          type: string
          nullable: true
        businessContext:
          type: string
          nullable: true
        versions:
          type: array
          items:
            $ref: "#/components/schemas/MeetingVersion"
        approvalHistory:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRecord"
        notes:
          type: array
          items:
            $ref: "#/components/schemas/MeetingNote"
        decisions:
          type: array
          items:
            $ref: "#/components/schemas/DecisionItem"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/ActionItem"
        risks:
          type: array
          items:
            $ref: "#/components/schemas/RiskItem"
        commitments:
          type: array
          items:
            $ref: "#/components/schemas/CommitmentItem"
        qualityFlags:
          type: array
          items:
            type: string
        partial:
          type: boolean
          description: True when some panels failed to load
...